services/trade_to_ohlc/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["max-mcd <max-mcd@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
quixstreams = "^2.6.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
